feat: Standard C⋆-module (#7)
# The standard C⋆-module

Given a family `E : ι → Type*` of C⋆-modules over a C⋆-algebra `A`, the *standard C⋆-module* `ℓ²(A, E)`
consists of the additive subgroup of `Π i, E i` consisting of those `x : Π i, E i` such that the sum
`∑' i, ⟪x i, x i⟫_A` converges. Note that such convergence is a consequence of, but not equivalent to,
the convergence of `∑' i, ‖x i‖ ^ 2`. Because of the similarity with `lp` for `p = 2`, we develop the
API in an analogous manner.
`ℓ²(A, E)` is naturally a complete C⋆-module over `A`, with `A`-valued inner product given by
`⟪x, y⟫_A = ∑' i, ⟪x i, y i⟫_A`.

## Main declarations

+ `CStarModule.MemStandard`: The property on `x : Π i, E i` saying that `∑' i, ⟪x i, x i⟫_A` converges.
+ `CStarModule.Standard`: The standard C⋆-module `ℓ²(A, E)` which is the additive subgroup with carrier
  `{x : Π i, E i | MemStandard A x}`.
+ `CStarAlgebra.dominated_convergence`: If `x y : ι → A` are sequences of nonnegative elements with `x`
  summable and `y` dominated by `x`, then `y` is also summable.
+ `CStarModule.Standard.inst`: The standard C⋆-module `ℓ²(A, E)` is a C⋆-module over `A`.
+ `CStarModule.Standard.instCompleteSpace`: The standard C⋆-module `ℓ²(A, E)` is a complete space.